my_stock_project/requirements.txt
agate==1.9.1
alembic==1.17.2
altair==6.0.0
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.12.0
attrs==25.4.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==14.0
curl_cffi==0.14.0
daff==1.4.2
dagster==1.12.7
dagster-dbt==0.28.7
dagster-graphql==1.12.7
dagster-pipes==1.12.7
dagster-webserver==1.12.7
dagster_shared==1.12.7
dbt-adapters==1.22.1
dbt-common==1.37.2
dbt-core==1.10.16
dbt-extractor==0.6.0
dbt-postgres==1.9.1
dbt-protos==1.0.410
dbt-semantic-interfaces==0.9.0
deepdiff==8.6.1
docstring_parser==0.17.0
filelock==3.20.1
frozendict==2.4.7
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
gql==4.0.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-health-checking==1.76.0
h11==0.16.0
httptools==0.7.1
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
isodate==0.7.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
leather==0.4.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mashumaro==3.14
mdurl==0.1.2
more-itertools==10.8.0
msgpack==1.1.2
multidict==6.7.0
multitasking==0.0.12
narwhals==2.14.0
networkx==3.6.1
numpy==2.3.5
orderly-set==5.5.0
orjson==3.11.5
packaging==25.0
pandas==2.3.3
parsedatetime==2.6
pathlib_abc==0.5.2
pathspec==0.12.1
peewee==3.18.3
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.0
psycopg2-binary==2.9.11
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
snowplow-tracker==1.1.0
soupsieve==2.8.1
SQLAlchemy==2.0.45
sqlglot==28.0.0
sqlglotrs==0.7.3
sqlparse==0.5.5
starlette==0.50.0
streamlit==1.52.2
structlog==25.5.0
tabulate==0.9.0
tenacity==9.1.2
text-unidecode==1.3
toml==0.10.2
tomli==2.3.0
tomlkit==0.13.3
toposort==1.10
tornado==6.5.4
tqdm==4.67.1
typer==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
universal_pathlib==0.3.7
urllib3==2.6.2
uvicorn==0.40.0
watchdog==6.0.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
yfinance==0.2.66
zipp==3.23.0
